refactor AssertionSyntaxBuilder
